fix: refactor(parser): move condition parser wrapper to application core and mark old parser artifacts archived
